# raise-openapi.yaml
openapi: 3.0.0
paths:
...
  /advertisements/create:
    post:
      operationId: create
      summary: creating advertisement either for internal or external (admin only)
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementCreateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: >-
                            #/components/schemas/AdvertisementCreateCommandResult
        '201':
          description: ''
      tags: &ref_13
        - AdvertisementModule
  /advertisements:
    get:
      operationId: findMany
      summary: find advertisement either for internal or external (admin only)
      parameters:
        - name: page
          required: false
          in: query
          description: Page
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Limit
          schema:
            default: 10
            type: number
        - name: sorting_field
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdvertisementEntity'
      tags: *ref_13
  /advertisements/mine:
    get:
      operationId: findMyAdvertisement
      summary: find personalized ads for curtain user
      parameters:
        - name: page
          required: false
          in: query
          description: Page
          schema:
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Limit
          schema:
            default: 10
            type: number
        - name: sorting_field
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdvertisementEntity'
      tags: *ref_13
  /advertisements/update:
    patch:
      operationId: update
      summary: updating advertisement either for internal or external
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdvertisementUpdateDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/AdvertisementUpdateCommandResult'
      tags: *ref_13
...
info:
  title: Tmra Raise
  description: Tmra Raise API
  version: '3.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    BaseResponse:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        data:
          type: object
        meta:
          type: object
      required:
        - statusCode
        - message
        - data
        - meta
    BeneficiaryEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        is_deleted:
          type: boolean
      required:
        - id
        - name
        - is_deleted
    ChequeEntity:
      type: object
      properties:
        id:
          type: string
        payment_id:
          type: string
        deposit_date:
          format: date-time
          type: string
        number:
          type: string
        transfer_receipt:
          type: object
        payment:
          $ref: '#/components/schemas/ProposalPaymentEntity'
      required:
        - id
        - payment_id
        - deposit_date
        - number
        - transfer_receipt
    PaymentStatusEntity:
      type: object
      properties:
        id:
          type: string
        payments:
          $ref: '#/components/schemas/ProposalPaymentEntity'
      required:
        - id
    ProposalPaymentEntity:
      type: object
      properties:
        id:
          type: string
        payment_amount:
          type: number
          nullable: true
        proposal_id:
          type: string
        payment_date:
          format: date-time
          type: string
          nullable: true
        status:
          type: string
          nullable: true
          default: SET_BY_SUPERVISOR
        number_of_payments:
          type: number
          nullable: true
        order:
          type: number
          nullable: true
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.891Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.891Z
        cheques:
          type: array
          items:
            $ref: '#/components/schemas/ChequeEntity'
        proposal:
          $ref: '#/components/schemas/ProposalEntity'
        payment_status:
          $ref: '#/components/schemas/PaymentStatusEntity'
      required:
        - id
        - proposal_id
    ProposalProjectTimelineEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        start_date:
          format: date-time
          type: string
        end_date:
          format: date-time
          type: string
        proposal_id:
          type: string
        proposal:
          $ref: '#/components/schemas/ProposalEntity'
      required:
        - id
        - name
        - start_date
        - end_date
        - proposal_id
    ProposalItemBudgetEntity:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        explanation:
          type: string
        clause:
          type: string
        proposal_id:
          type: string
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.894Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.894Z
        proposal:
          $ref: '#/components/schemas/ProposalEntity'
      required:
        - id
        - amount
        - explanation
        - clause
        - proposal_id
        - proposal
...
    UserTypeEntity:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    ProposalLogEntity:
      type: object
      properties:
        id:
          type: string
        proposal_id:
          type: string
        created_at:
          format: date-time
          type: string
          default: 2023-07-20T08:21:55.894Z
        updated_at:
          format: date-time
          type: string
          default: 2023-07-20T08:21:55.894Z
        reviewer_id:
          type: string
          nullable: true
        state:
          type: string
        new_values:
          type: object
        old_values:
          type: object
        notes:
          type: string
          nullable: true
        action:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        user_role:
          type: string
          nullable: true
        response_time:
          type: number
          nullable: true
        reject_reason:
          type: string
          nullable: true
        reviewer:
          $ref: '#/components/schemas/UserEntity'
        proposal:
          $ref: '#/components/schemas/ProposalEntity'
        notification:
          type: array
          items:
            $ref: '#/components/schemas/NotificationEntity'
        user_type:
          $ref: '#/components/schemas/UserTypeEntity'
      required:
        - id
        - proposal_id
        - created_at
        - updated_at
        - state
        - proposal
    ProposalTrackSectionEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        budget:
          type: number
        section_id:
          type: string
        is_leaf:
          type: boolean
          default: false
        track_id:
          type: string
        is_deleted:
          type: boolean
          default: false
        track_section:
          $ref: '#/components/schemas/ProposalTrackSectionEntity'
        other_track_section:
          type: array
          items:
            $ref: '#/components/schemas/ProposalTrackSectionEntity'
        track:
          $ref: '#/components/schemas/TrackEntity'
      required:
        - id
        - name
        - budget
        - is_leaf
        - track_id
        - is_deleted
        - track
    TrackEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          nullable: true
        with_consultation:
          type: boolean
          nullable: true
          default: false
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.898Z
        is_deleted:
          type: boolean
          nullable: true
          default: false
        proposals:
          type: array
          items:
            $ref: '#/components/schemas/ProposalEntity'
        sections:
          type: array
          items:
            $ref: '#/components/schemas/ProposalTrackSectionEntity'
        user:
          type: array
          items:
            $ref: '#/components/schemas/UserEntity'
      required:
        - id
    ProposalEntity:
      type: object
      properties:
        id:
          type: string
        accreditation_type_id:
          type: string
          nullable: true
        added_value:
          type: string
          nullable: true
        amount_required_fsupport:
          type: number
          nullable: true
        been_made_before:
          type: boolean
          nullable: true
          default: false
        been_supported_before:
          type: boolean
          nullable: true
          default: false
        beneficiary_detail:
          $ref: '#/components/schemas/BeneficiaryEntity'
        beneficiary_id:
          type: string
          nullable: true
        cashier_id:
          type: string
          nullable: true
        chairman_of_board_of_directors:
          type: string
          nullable: true
        clasification_field:
          type: string
          nullable: true
        clause:
          type: string
          nullable: true
        closing_report:
          type: boolean
          nullable: true
          default: false
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.885Z
        does_an_agreement:
          type: boolean
          nullable: true
          default: false
        execution_time:
          type: number
          nullable: true
        finance_id:
          type: string
          nullable: true
        follow_ups:
          type: array
          items:
            $ref: '#/components/schemas/ProposalFollowUpEntity'
        fsupport_by_supervisor:
          type: number
          nullable: true
        governorate:
          type: string
          nullable: true
        inclu_or_exclu:
          type: boolean
          nullable: true
          default: false
        inner_status:
          type: string
          nullable: true
          default: CREATED_BY_CLIENT
        letter_ofsupport_req:
          type: object
        most_clents_projects:
          type: string
          nullable: true
        need_consultant:
          type: boolean
          nullable: true
        need_picture:
          type: boolean
          nullable: true
          default: false
        num_ofproject_binicficiaries:
          type: number
          nullable: true
        number_of_payments:
          type: number
          nullable: true
        number_of_payments_by_supervisor:
          type: number
          nullable: true
        oid:
          type: number
          nullable: true
        old_inner_status:
          type: string
          nullable: true
        on_consulting:
          type: boolean
          nullable: true
          default: false
        on_revision:
          type: boolean
          nullable: true
          default: false
        outter_status:
          type: string
          nullable: true
          default: ONGOING
        partial_support_amount:
          type: number
          nullable: true
        payments:
          type: array
          items:
            $ref: '#/components/schemas/ProposalPaymentEntity'
        pm_email:
          type: string
          nullable: true
        pm_mobile:
          type: string
          nullable: true
        pm_name:
          type: string
          nullable: true
        previously_add_bank:
          nullable: true
          type: array
          items:
            type: string
        project_attachments:
          type: object
        project_beneficiaries:
          type: string
          nullable: true
        project_beneficiaries_specific_type:
          type: string
          nullable: true
        project_goals:
          type: string
          nullable: true
        project_idea:
          type: string
          nullable: true
        project_implement_date:
          format: date-time
          type: string
          nullable: true
        project_location:
          type: string
          nullable: true
        project_manager_id:
          type: string
          nullable: true
        project_name:
          type: string
        project_number:
          type: number
          nullable: true
        project_numbers1:
          type: number
          nullable: true
        project_outputs:
          type: string
          nullable: true
        project_risks:
          type: string
          nullable: true
        project_strengths:
          type: string
          nullable: true
        project_timeline:
          type: array
          items:
            $ref: '#/components/schemas/ProposalProjectTimelineEntity'
        project_track:
          type: string
          nullable: true
        proposal_bank_id:
          type: string
          nullable: true
        proposal_beneficiaries:
          type: array
          items:
            $ref: '#/components/schemas/BeneficiaryEntity'
        proposal_item_budgets:
          type: array
          items:
            $ref: '#/components/schemas/ProposalItemBudgetEntity'
        proposal_log:
          type: array
          items:
            $ref: '#/components/schemas/ProposalLogEntity'
        reasons_to_accept:
          type: string
          nullable: true
        region:
          type: string
          nullable: true
        remote_or_insite:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
          default: MODERATOR
        step:
          type: string
          nullable: true
        submitter_user_id:
          type: string
        supervisor_id:
          type: string
          nullable: true
        support_goal_id:
          type: string
          nullable: true
        support_outputs:
          type: string
          nullable: true
        support_type:
          type: boolean
          nullable: true
          default: false
        target_group_age:
          type: string
          nullable: true
        target_group_num:
          type: number
          nullable: true
        target_group_type:
          type: string
          nullable: true
        track:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/TrackEntity'
        track_id:
          type: string
          nullable: true
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.885Z
        vat:
          type: boolean
          nullable: true
          default: false
        vat_percentage:
          type: number
          nullable: true
        whole_budget:
          type: number
          nullable: true
        user:
          $ref: '#/components/schemas/UserEntity'
        supervisor:
          $ref: '#/components/schemas/UserEntity'
      required:
        - id
        - letter_ofsupport_req
        - project_attachments
        - project_name
        - submitter_user_id
    BankInformationEntity:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        bank_account_name:
          type: string
          nullable: true
        bank_account_number:
          type: string
          nullable: true
        bank_name:
          type: string
          nullable: true
        proposal_id:
          type: string
          nullable: true
        card_image:
          type: object
        is_deleted:
          type: boolean
          nullable: true
        bank_id:
          type: string
          nullable: true
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.888Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.888Z
        user:
          $ref: '#/components/schemas/UserEntity'
        proposal_bank_informationToproposal_proposal_bank_id:
          $ref: '#/components/schemas/ProposalEntity'
      required:
        - id
        - user_id
        - user
    UserRoleEntity:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        user_type_id:
          type: string
      required:
        - id
        - user_id
        - user_type_id
    FileManagerEntity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        size:
          type: number
        mimetype:
          type: string
        url:
          type: string
        table_name:
          type: string
          nullable: true
        column_name:
          type: string
          nullable: true
        user_id:
          type: string
          nullable: true
        is_deleted:
          type: boolean
          default: false
        created_at:
          format: date-time
          type: string
          default: 2023-07-20T08:21:55.900Z
        updated_at:
          format: date-time
          type: string
          default: 2023-07-20T08:21:55.900Z
        proposal_id:
          type: string
          nullable: true
        bank_information_id:
          type: string
          nullable: true
        advertisement_id:
          type: string
          nullable: true
        fid:
          type: number
          nullable: true
        user:
          $ref: '#/components/schemas/UserEntity'
      required:
        - id
        - name
        - size
        - mimetype
        - url
        - is_deleted
        - created_at
        - updated_at
...
    UserStatusEntity:
      type: object
      properties:
        id:
          type: string
        client_log:
          type: object
        user:
          type: array
          items:
            $ref: '#/components/schemas/UserEntity'
        user_status_log:
          type: array
          items:
            type: object
      required:
        - id
        - client_log
    UserEntity:
      type: object
      properties:
        id:
          type: string
        employee_name:
          type: string
          nullable: true
        mobile_number:
          type: string
          nullable: true
        email:
          type: string
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.887Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.887Z
        employee_path:
          type: string
          nullable: true
        last_login:
          format: date-time
          type: string
          nullable: true
          default: null
        status_id:
          type: string
        address:
          type: string
          nullable: true
          default: null
        google_session:
          type: object
        is_online:
          type: boolean
          nullable: true
          default: null
        track_id:
          type: string
          nullable: true
          default: null
        uid:
          type: number
          nullable: true
          default: null
        bank_information:
          type: array
          items:
            $ref: '#/components/schemas/BankInformationEntity'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/UserRoleEntity'
        file_manager:
          type: array
          items:
            $ref: '#/components/schemas/FileManagerEntity'
        track:
          $ref: '#/components/schemas/TrackEntity'
        proposal_proposal_project_manager_idTouser:
          type: array
          items:
            $ref: '#/components/schemas/ProposalEntity'
        proposals:
          type: array
          items:
            $ref: '#/components/schemas/ProposalEntity'
        proposal_proposal_supervisor_idTouser:
          type: array
          items:
            $ref: '#/components/schemas/ProposalEntity'
        client_data:
          $ref: '#/components/schemas/ClientDataEntity'
        status:
          $ref: '#/components/schemas/UserStatusEntity'
      required:
        - id
        - email
        - status_id
        - google_session
        - roles
        - status
    ProposalFollowUpEntity:
      type: object
      properties:
        id:
          type: string
        proposal_id:
          type: string
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.887Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.887Z
        attachments:
          type: object
        content:
          type: string
          nullable: true
        user_id:
          type: string
        submitter_role:
          type: string
        employee_only:
          type: boolean
        user:
          $ref: '#/components/schemas/UserEntity'
        proposal:
          $ref: '#/components/schemas/ProposalEntity'
      required:
        - id
        - proposal_id
        - user_id
        - submitter_role
        - employee_only
...
    AdvertisementEntity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        title:
          type: string
        content:
          type: string
        track_id:
          type: string
          nullable: true
        track:
          $ref: '#/components/schemas/TrackEntity'
        logo:
          type: object
        date:
          format: date-time
          type: string
        start_time:
          type: string
        created_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.932Z
        updated_at:
          format: date-time
          type: string
          nullable: true
          default: 2023-07-20T08:21:55.932Z
      required:
        - id
        - type
        - title
        - content
        - date
        - start_time
        - created_at
        - updated_at
    AdvertisementCreateCommandResult:
      type: object
      properties:
        advertisement:
          $ref: '#/components/schemas/AdvertisementEntity'
        created_file_managers:
          type: array
          items:
            type: string
      required:
        - advertisement
        - created_file_managers
    AdvertisementCreateDto:
      type: object
      properties:
        content:
          type: string
        title:
          type: string
        track_id:
          type: string
        type:
          enum:
            - INTERNAL
            - EXTERNAL
          type: string
          description: description of the AdvertisementTypeEnum property
        date:
          type: string
        start_time:
          type: string
      required:
        - content
        - title
        - type
        - date
        - start_time
    PaginatedResponse:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        data:
          type: object
        meta:
          type: object
        totalData:
          type: number
          description: Total number of documents in collection that match a query {Number}
          default: 0
        limit:
          type: number
          description: Limit that was used {Number}
          default: 10
        page:
          type: number
          description: Current page number {Number}
          default: 1
        totalPages:
          type: number
          description: Total number of pages. {Number}
          default: 1
        pagingCounter:
          type: number
          description: >-
            The starting index/serial/chronological number of first document in
            current page. {Number}

            (Eg: if page=2 and limit=10, then pagingCounter will be 11)
          default: 1
        hasPrevPage:
          type: boolean
          description: Availability of prev page. {Boolean}
          default: false
        hasNextPage:
          type: boolean
          description: Availability of next page. {Boolean}
          default: false
        prevPage:
          type: object
          description: Previous page number if available or NULL. {Number}
          default: null
        nextPage:
          type: object
          description: Next page number if available or NULL. {Number}
          default: null
        offset:
          type: object
          description: Only if specified or default page/offset values were used {Number}
      required:
        - statusCode
        - message
        - data
        - meta
    AdvertisementUpdateCommandResult:
      type: object
      properties: {}
    AdvertisementUpdateDto:
      type: object
      properties:
        advertisement_id:
          type: string
        content:
          type: string
        title:
          type: string
        track_id:
          type: string
        type:
          enum:
            - INTERNAL
            - EXTERNAL
          type: string
        date:
          type: string
        start_time:
          type: string
        deleted_logo_urls:
          type: array
          items:
            type: string
      required:
        - advertisement_id